[REF] mail: subchannel feature use channel
