+++ addons/mail/static/src/discuss/core/public_web/message_actions.js
@@ -4,14 +4,14 @@ import { _t } from "@web/core/l10n/translation";
 registerMessageAction("create-or-view-thread", {
     condition: ({ message, store, thread }) =>
         message.thread?.eq(thread) &&
-        message.thread.hasSubChannelFeature &&
+        message.channel_id?.hasSubChannelFeature &&
         store.self_user?.share === false,
     icon: "fa fa-comments-o",
     onSelected: ({ message }) => {
         if (message.linkedSubChannel) {
             message.linkedSubChannel.open({ focus: true });
         } else {
-            message.thread.createSubChannel({ initialMessage: message });
+            message.channel_id?.createSubChannel({ initialMessage: message });
         }
     },
     name: ({ message }) => (message.linkedSubChannel ? _t("View Thread") : _t("Create Thread")),
